+++ relocate.js
@@ -1,6 +1,4 @@
 const fs = require("fs");
-const path = require("path");
-const glob = require("glob");
 
 fs.copyFile("./dist/core/env.js", "./env.js", error => {
     if (error) {