+++ src/core/types.d.ts
@@ -1,7 +1,1 @@
-export type WorkerResponse = {
-    type: "load";
-    files: Array<string>;
-    requestUid: string | null;
-};
-
 export type WebComponentLoad = null | "lazy" | "eager";